NN-4787 refactor to next steps
