+++ server/routes/nextSteps/police/index.ts
@@ -1,10 +1,10 @@
 import express, { RequestHandler, Router } from 'express'
 import ProsecutionRoutes from './prosecution'
-import asyncMiddleware from '../../middleware/asyncMiddleware'
+import asyncMiddleware from '../../../middleware/asyncMiddleware'
 
-import adjudicationUrls from '../../utils/urlGenerator'
-import UserService from '../../services/userService'
-import OutcomesService from '../../services/outcomesService'
+import adjudicationUrls from '../../../utils/urlGenerator'
+import UserService from '../../../services/userService'
+import OutcomesService from '../../../services/outcomesService'
 
 export default function prosecutionRoutes({
   userService,
@@ -20,8 +20,8 @@ export default function prosecutionRoutes({
   const get = (path: string, handler: RequestHandler) => router.get(path, asyncMiddleware(handler))
   const post = (path: string, handler: RequestHandler) => router.post(path, asyncMiddleware(handler))
 
-  get(adjudicationUrls.prosecution.matchers.start, prosecutionRoute.view)
-  post(adjudicationUrls.prosecution.matchers.start, prosecutionRoute.submit)
+  get(adjudicationUrls.nextStepsPolice.matchers.start, prosecutionRoute.view)
+  post(adjudicationUrls.nextStepsPolice.matchers.start, prosecutionRoute.submit)
 
   return router
 }
