+++ server/routes/nextSteps/police/prosecution.test.ts
@@ -1,12 +1,12 @@
 import { Express } from 'express'
 import request from 'supertest'
-import appWithAllRoutes from '../testutils/appSetup'
-import adjudicationUrls from '../../utils/urlGenerator'
-import UserService from '../../services/userService'
-import OutcomesService from '../../services/outcomesService'
+import appWithAllRoutes from '../../testutils/appSetup'
+import adjudicationUrls from '../../../utils/urlGenerator'
+import UserService from '../../../services/userService'
+import OutcomesService from '../../../services/outcomesService'
 
-jest.mock('../../services/userService')
-jest.mock('../../services/outcomesService')
+jest.mock('../../../services/userService')
+jest.mock('../../../services/outcomesService')
 
 const userService = new UserService(null) as jest.Mocked<UserService>
 const outcomesService = new OutcomesService(null) as jest.Mocked<OutcomesService>
@@ -22,14 +22,14 @@ afterEach(() => {
   jest.resetAllMocks()
 })
 
-describe('GET /prosecution', () => {
+describe('GET /nextSteps/police', () => {
   beforeEach(() => {
     app = appWithAllRoutes({ production: false }, { userService, outcomesService }, {})
     userService.getUserRoles.mockResolvedValue(['NOT_REVIEWER'])
   })
   it('should load the `Page not found` page', () => {
     return request(app)
-      .get(adjudicationUrls.prosecution.urls.start(100))
+      .get(adjudicationUrls.nextStepsPolice.urls.start(100))
       .expect('Content-Type', /html/)
       .expect(res => {
         expect(res.text).toContain('Page not found')
@@ -37,10 +37,10 @@ describe('GET /prosecution', () => {
   })
 })
 
-describe('GET /prosecution', () => {
+describe('GET /nextSteps/police', () => {
   it('should load the `Prosecution` page', () => {
     return request(app)
-      .get(adjudicationUrls.prosecution.urls.start(100))
+      .get(adjudicationUrls.nextStepsPolice.urls.start(100))
       .expect('Content-Type', /html/)
       .expect(res => {
         expect(res.text).toContain('Will this charge continue to prosecution?')
@@ -48,10 +48,10 @@ describe('GET /prosecution', () => {
   })
 })
 
-describe('POST /prosecution', () => {
+describe('POST /nextSteps/police', () => {
   it('should successfully call the endpoint and redirect', () => {
     return request(app)
-      .post(`${adjudicationUrls.prosecution.urls.start(100)}`)
+      .post(`${adjudicationUrls.nextStepsPolice.urls.start(100)}`)
       .send({
         prosecutionChosen: 'yes',
         nextStepChosen: null,
